+++ ChangeLog.rst
@@ -13,6 +13,12 @@ Note worthy changes
 
 - Added new socialaccount provider: Klaviyo.
 
+Fixes
+-----
+
+- BitBucket: When using the BitBucket API, the token is passed in the headers
+  instead of the query parameters, which no longer works since May 4th, 2026
+  (`deprecation notice <https://developer.atlassian.com/cloud/bitbucket/changelog/#CHANGE-3052>`__).
 
 65.16.1 (2026-04-17)
 ********************